--- app/src/main/java/woowacourse/payments/domain/ExpireDateValidationException.kt
@@ -1,7 +1,0 @@
-package woowacourse.payments.domain
-
-import woowacourse.payments.domain.ExpireDateStatus.Invalid.ExpireDateInvalidReason
-
-class ExpireDateValidationException(
-    val reason: ExpireDateInvalidReason,
-) : IllegalArgumentException()
