--- app/src/main/java/woowacourse/payments/domain/PaymentCard.kt
@@ -1,8 +1,0 @@
-package woowacourse.payments.domain
-
-data class PaymentCard(
-    val cardNumber: CardNumber,
-    val expireDate: ExpireDate,
-    val ownerName: OwnerName,
-    val password: Password,
-)
