+++ app/src/main/java/woowacourse/payments/domain/card/exception/CardNumberException.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.domain.exception
+package woowacourse.payments.domain.card.exception
 
 sealed class CardNumberException : IllegalArgumentException() {
     data object InvalidLength : CardNumberException()
