+++ app/src/main/java/woowacourse/payments/domain/card/exception/ExpireDateException.kt
@@ -1,0 +1,7 @@
+package woowacourse.payments.domain.card.exception
+
+import woowacourse.payments.domain.card.ExpireDateStatus
+
+class ExpireDateException(
+    val reason: ExpireDateStatus.Invalid.ExpireDateInvalidReason,
+) : IllegalArgumentException()
