+++ app/src/main/java/woowacourse/payments/domain/card/exception/OwnerNamerException.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.domain.exception
+package woowacourse.payments.domain.card.exception
 
 sealed class OwnerNamerException : IllegalArgumentException() {
     data object InvalidLength : OwnerNamerException()
