+++ app/src/main/java/woowacourse/payments/domain/card/values/ExpireDate.kt
@@ -1,6 +1,7 @@
-package woowacourse.payments.domain
+package woowacourse.payments.domain.card.values
 
-import woowacourse.payments.domain.ExpireDateStatus.Invalid.ExpireDateInvalidReason
+import woowacourse.payments.domain.card.ExpireDateStatus.Invalid.ExpireDateInvalidReason
+import woowacourse.payments.domain.card.exception.ExpireDateException
 import java.time.YearMonth
 
 @JvmInline
@@ -9,7 +10,7 @@ value class ExpireDate private constructor(
 ) {
     init {
         if (value < YearMonth.now()) {
-            throw ExpireDateValidationException(ExpireDateInvalidReason.EXPIRED)
+            throw ExpireDateException(ExpireDateInvalidReason.EXPIRED)
         }
     }
 
@@ -20,21 +21,21 @@ value class ExpireDate private constructor(
 
         fun from(expireDateString: String): Result<ExpireDate> {
             if (expireDateString.length != MAX_LENGTH_EXPIRE_DATE) {
-                return Result.failure(ExpireDateValidationException(ExpireDateInvalidReason.INVALID_FORMAT))
+                return Result.failure(ExpireDateException(ExpireDateInvalidReason.INVALID_FORMAT))
             }
 
             val mm =
                 expireDateString.take(2).toIntOrNull() ?: return Result.failure(
-                    ExpireDateValidationException(ExpireDateInvalidReason.INVALID_FORMAT),
+                    ExpireDateException(ExpireDateInvalidReason.INVALID_FORMAT),
                 )
 
             val yy =
                 expireDateString.takeLast(2).toIntOrNull() ?: return Result.failure(
-                    ExpireDateValidationException(ExpireDateInvalidReason.INVALID_FORMAT),
+                    ExpireDateException(ExpireDateInvalidReason.INVALID_FORMAT),
                 )
 
             if (mm !in 1..12) {
-                return Result.failure(ExpireDateValidationException(ExpireDateInvalidReason.INVALID_MONTH))
+                return Result.failure(ExpireDateException(ExpireDateInvalidReason.INVALID_MONTH))
             }
 
             val yearMonth = YearMonth.of(2000 + yy, mm)
