+++ app/src/main/java/woowacourse/payments/ui/components/PaymentCardPlate.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.em
 import androidx.compose.ui.unit.sp
 import woowacourse.payments.R
-import woowacourse.payments.domain.CardNumber
+import woowacourse.payments.domain.card.values.CardNumber
 import woowacourse.payments.ui.model.PaymentCardUiModel
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
