+++ app/src/main/java/woowacourse/payments/ui/features/addcard/components/CardPasswordField.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import woowacourse.payments.R
-import woowacourse.payments.domain.Password.Companion.MAX_LENGTH_PASSWORD
+import woowacourse.payments.domain.card.values.Password.Companion.MAX_LENGTH_PASSWORD
 import woowacourse.payments.ui.components.AppTextField
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
